pyproject.toml
[project]
name = "nmde"
version = "0.1.0"
description = "A collection of tools for the NMDE environment."
authors = [{ name = "Gemini", email = "gemini@google.com" }]
dependencies = [
    "docker>=7.1.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "textual>=5.3.0",
]
requires-python = ">=3.10"

[project.scripts]
nmde = "nmde.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nmde"]
